Use indexed data
Instead of always getting all of the data for a search, only data that could match the query is retrieved. For example, if the query starts with "K", usually for Kepler, only data for exoplanets that have host stars starting with the letter K is downloaded. Similarily with the discovery year, "2010" will only get the data for exoplanets discovered in 2010. This results in small improvements to download size and potential increases in speed, at the cost of having a somewhat larger "database" with all of the different JSON files.